API/requirements.txt
fastapi
pydantic
numpy
tensorflow
torch
torchvision
scikit-learn
joblib
pillow
uvicorn
